refactor(font-mapping): Flatten the hierarchy as the package is not that big.
