+++ packages/ckeditor5-font-mapper/src/ConfigToMapUtil.ts
