+++ packages/ckeditor5-font-mapper/src/FontMapping.ts
