+++ packages/ckeditor5-font-mapper/src/FontMappingRegistry.ts
