+++ packages/ckeditor5-font-mapper/src/FontReplacer.ts
