+++ packages/ckeditor5-font-mapper/src/HtmlEncodingMap.ts
