+++ packages/ckeditor5-font-mapper/src/SymbolFontMap.ts
