--- packages/ckeditor5-font-mapper/src/fontMapping/index-doc.ts
@@ -1,7 +1,0 @@
-/**
- * @packageDocumentation
- * @category Virtual
- */
-export * from "./FontMapping";
-export * from "./FontMappingRegistry";
-export * from "./SymbolFontMap";
